ユーザ変更画面の処理の変更

+++ src/main/java/jp/ac/chitose/ir/application/service/management/UsersData.java
@@ -1,7 +1,7 @@
 package jp.ac.chitose.ir.application.service.management;
 
 public record UsersData(
-        int id,
+        long id,
         String login_id,
         String user_name,
         String password,
